refactor: Change background style correct spelling
This commit will change the way the gradient background is added to
the site. Since it is planned to use and inverted background for the
contacts section, the background for the main part is moved to main.
Some width/heigh and margin adjustments have been necessary. A spelling
error in the motivational text has been corrected.

- Move gradient background to main section
- Adjust height/ width and margin
- Correct spelling error in motivational text