feat: Add more details to quiz html construct
This commit will add more details to the quiz html construct and
should form the final state of it. However, since there are still
many open questions on how to implement exaclty with JS, it might
change again. The main purpose should be to style the construct
with CSS to have a start base.