feat: Add structure/style on quiz & fix index site
This commit will refactor the html structure of the quit site and
add finally style and position of the main elements. The quiz card
is centered and placeholders for the content are added.

It will also fix some class reference on the index page to avoid
targeting direct html elements via css. This is important since I
use only one css file, so there are multiple "main" html elements,
hence it is better to target via class attribute.

- Improve html structure of quiz site
- Add basic form & style to match the other sites
- Center the quiz card and style it
- Add content placeholder for quiz card
- Correct index site css targeting of html elements