feat: Add & display random question
This commit will enhance the function to pick a random question out
of the question pool. It will also set the index for the current
question, which is needed to check the total amount, when creating
the counter later on.
It will then write the question back to the DOM