feat: Enhance display question function
This commit will enhance the display question function and
create a button element with the answers out of the answers array.
It will also set a dataset correct element to the correct answer.
Button child elements are created and appended to the button
elements.
An event listener on the click event is created and it will handover
the click event to the chooseAnswer function.