feat: Add function to clear answer elements
This commit will add the function to clear the already visible
answer buttons, the one from the previous question.
It will clear the first child element of the answButtonsElement, since
the new answers are the 2nd child element.

The next question button will also be hidden, since we now need to
select an answer.