feat: Enhance checkAnswer & add set/clear state
This commit will enhance the checkAnswer function and set the button
dataset to the correct value. It will then unhide the next button
in order to be able to press it.

The setState function will be called which first calls the
clearState function, which will clear the css class for each
element (correct or wrong).
It will afterwards evaluate the value in button dataset and if
correct it will set css class correct, else wrong.