feat: Add check for available questions
This commit will enhance the chooseAnswer function to check if the
curr question index is < than question pool. If so, next question will
be displayed and the curr question index will be increased by 1.
Else the end of the game is reached and the next button will change
to restart button. Note, this is only temporary, it will later be
replaced with a final message prompt.