feat: Add html and css part of quiz for JS part
This commit will bring in the html and css part, which is used
for the quiz.js file. It adds all id slectors in the html file
and changes slightly the html structure. For the question buttons
grid is used to have them using the available space. The rest
keeps flexbox.

In css some classes will be added to transform the color of the
buttons to either green or red. These classes will be targeted via
the js function.