feat: Add question counter feature
This commit will add the question counter feature to the game.
It will use the currenct question index and add 1 to it (since the
index in the array starts at 0). It will count the total amount
of questions via checking the length of the array.
It will write both parameters to the dom.