refactor: Adopt design of index, confirmation site
This commit will change the index site with most of its content to
reflect the current state of the development phase. The
unnecessary buttons have been removed and the gradient effect adjusted.

The flexbox construct is now aligned for all sites (except quz.html
which will be merged into index.html in the next step).
The Flexbox construct follows now the grad-grandperent, grandparent,
parent & child approach, which will make sure to have all cards
centered on x & y axis and fully responsive.

The construct is done in order to be able to simply hide or unhide
a card and replace it with another card. This is important for the
javascript game construct.