refactor: Integrate quiz into index.html site
This commit will merge the quiz.html into index.html. The quiz
has initially been developed in a seperate site, which turned out
to be a bad decision regarding performance, clean code and
functionality.
Therefore this commit is the first step to integrate the quiz into
the index.html site. The necessary adoptions on css (some flexbox
adjustments) are also included.
Further changes will come to complete this merge.

The JS logic is adopted to reflext this change. The startGame
button is moved to the game-card and a new restart button is
introduced, once the end of the questins is reached.