feat: Function to randomize questions out of pool
This commit will add many questions to the questions.js file, taken
from the opentdb via the Ruby script.
It will also randomize the order of the questions every time
the question database will be "polled". This commit will also
add the functionality that only a certain amount of questions gets
pulled from the database.
The commit will also change the order of which the js files are
getting read on the index.html file.